Improve output/interface
Now the layout of the output is similar to that of reghdfe. This includes a preamble with the number of observations and the partial F stat. There is also a standard error , t-value estimate, and the corresponding p-value. I did not adjust the output of the counter "s". We may also want to add R-squared and a note saying how the CI was determined at a later update, to further match other function outputs.