Add CI options
This update allows users to specify how they want the CI to be estimated AR (default), bootstrap, or asymptotic. This is done by the vce command. reps specifies the number of repetitions for the bootstrapping. Note that the interface does not yet give warnings, and it is not completely standardized across the CI choice.  The "first" feature is also incomplete.