adds display aiv; uses radius closest to zero to estimate SE from AR CI
displayaiv is now an option to show the coefficient on the estimated values of the anti-instrumental variable. This is only computed if the vce is set to asymptotic or to bootstrap, so this feature is unavailable in the Anderson-Rubin CI case. Also, in the Anderson-Rubin CI case, because the CI is uncentered, we are now inferring SE from the radius of the confidence interval closest to zero, instead of averaging the two radiuses, as this is more appropriate for hypothesis testing against zero.